+++ .env
@@ -1,3 +1,3 @@
-VUE_APP_SERVER_URL=https://api.splatnet2.com
+VUE_APP_SERVER_URL=http://localhost:3030
 VUE_APP_SERVER_API_VER=v1
-VUE_APP_VERSION=1.0.3.6
+VUE_APP_VERSION=1.0.5
